UI Freq Control / N3FJP Interface
  * Adjustments to the N3FJP interface code related to FC changes
  * Additional ULL data type-related changes to log-related functionality

+++ src/logbook/fd_logger.cxx
@@ -381,7 +381,7 @@ static std::string FD_opband()
 {
 	if (!active_modem) return "40";
 
-	float freq = qsoFreqDisp->value();
+	double freq = qsoFreqDisp->value();
 	freq /= 1e6;
 
 	if (freq >= 1.8 && freq < 3.5) return "160";
