+++ src/misc/macros.cxx
@@ -1727,8 +1727,8 @@ static void pQSYPLUS(std::string &s, size_t &i, size_t endbracket)
 		substitute(s, i, endbracket, "");
 		return;
 	}
-	int rf = 0;
-	float rfd = 0;
+	unsigned long long rf = 0;
+	double rfd = 0;
 	std::string sIncrFreq = s.substr(i+6, endbracket - i - 6);
 	// no frequency(s) specified
 	if (sIncrFreq.length() == 0) {
@@ -1736,9 +1736,9 @@ static void pQSYPLUS(std::string &s, size_t &i, size_t endbracket)
 		return;
 	}
 	// rf first value
-	sscanf(sIncrFreq.c_str(), "%f", &rfd);
+	sscanf(sIncrFreq.c_str(), "%lf", &rfd);
 	if (rfd != 0) {
-		rf = wf->rfcarrier() + (int)(1000*rfd);
+		rf = wf->rfcarrier() + (1000.0*rfd);
 		qsy(rf, active_modem ? active_modem->get_freq() : 1500);
 	}
 	substitute(s, i, endbracket, "");
@@ -3661,7 +3661,7 @@ static void doQSY(std::string s)
 	// rf first value
 	sscanf(sGoFreq.c_str(), "%lf", &rfd);
 	if (rfd > 0)
-		rf = (unsigned long long)(1000*rfd);
+		rf = (unsigned long long)(1000.0*rfd);
 	size_t pos;
 	if ((pos = sGoFreq.find(":")) != std::string::npos) {
 		// af second value
@@ -4135,15 +4135,15 @@ void set_macro_env(void)
 
 	// frequencies
 	char dial_freq[20];
-	snprintf(dial_freq, sizeof(dial_freq), "%ld", (long)wf->rfcarrier());
+	snprintf(dial_freq, sizeof(dial_freq), "%llu", wf->rfcarrier());
 	env[FLDIGI_DIAL_FREQUENCY].val = dial_freq;
 	char audio_freq[6];
 	snprintf(audio_freq, sizeof(audio_freq), "%d", active_modem->get_freq());
 	env[FLDIGI_AUDIO_FREQUENCY].val = audio_freq;
 	char freq[20];
-	snprintf(freq, sizeof(freq), "%ld", (long)(wf->rfcarrier() + (wf->USB()
-																  ? active_modem->get_freq()
-																  : -active_modem->get_freq())));
+	snprintf(freq, sizeof(freq), "%llu", (wf->rfcarrier() + (wf->USB()
+															  ? active_modem->get_freq()
+															  : -active_modem->get_freq())));
 	env[FLDIGI_FREQUENCY].val = freq;
 
 	// debugging vars
